+++ src/applications/edu-benefits/10215/tests/helper.unit.spec.jsx
@@ -89,6 +89,7 @@ describe('getFTECalcs', () => {
       supportedFTEPercent: null,
     });
   });
+
   it('should render the pdf download link', () => {
     const downloadLink = 'Download Link';
     const goBack = () => {};
