+++ src/applications/edu-benefits/10216/tests/10216-edu-benefits.cypress.spec.js
@@ -7,7 +7,7 @@ import maximalJson from './fixtures/data/maximal-test.json';
 import formConfig from '../config/form';
 import manifest from '../manifest.json';
 
-import { daysAgoYyyyMmDd } from '../../10215/helpers';
+import { daysAgoYyyyMmDd } from '../../utils/helpers';
 
 const mockManifest = {
   appName: manifest.appName,
